// app/components/GmMessage/type.ts
export type GmMessageProps = {
    banner: string;
    bannerAlt: string;
    pageTitle: string;
    metaTitle: string;
    metaDescription: string;
    firstSection: {
        title: string;
        firstDescription: string;
        secondDescription: string;
        image: string;
        imageAlt: string;
        name: string;
        designation: string;
    }
}